move/remove/edit cases
