+++ howso/client/base.py
@@ -1,22 +1,23 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from collections.abc import Collection, Iterable, Mapping, MutableMapping
+from collections.abc import Collection, Mapping, MutableMapping
 import typing as t
 from uuid import UUID
 import warnings
 
 from pandas import DataFrame, Index
 
-from howso.utilities import utilities as util
 from howso.utilities import internals
+from howso.utilities import utilities as util
+from howso.utilities.features import serialize_cases
 from .exceptions import HowsoError
 
 if t.TYPE_CHECKING:
     from howso.client.schemas import HowsoVersion, Project, Reaction, Session, Trainee, TraineePersistence
     from .cache import TraineeCache
     from .configuration import HowsoConfiguration
-    from .typing import CaseIndices, Cases
+    from .typing import CaseIndices, Cases, Precision
 
 
 class AbstractHowsoClient(ABC):
@@ -125,6 +126,10 @@ class AbstractHowsoClient(ABC):
         int
             The response payload size.
         """
+
+    def _resolve_trainee_id(self, trainee_id: str, *args, **kwargs):
+        """Resolve trainee identifier."""
+        return trainee_id
 
     @abstractmethod
     def get_version(self) -> HowsoVersion:
@@ -264,13 +269,14 @@ class AbstractHowsoClient(ABC):
             might be slower. Higher values should improve performance but may
             decrease accuracy of results.
         """
-        self._auto_resolve_trainee(trainee_id)
         if not self.active_session:
             raise HowsoError(self.ERROR_MESSAGES["missing_session"], code="missing_session")
         util.validate_list_shape(features, 1, "features", "str")
         util.validate_list_shape(features_to_impute, 1, "features_to_impute", "str")
+
         if self.configuration.verbose:
             print(f'Imputing Trainee with id: {trainee_id}')
+        self._auto_resolve_trainee(trainee_id)
         self._execute(trainee_id, "impute", {
             "features": features,
             "features_to_impute": features_to_impute,
@@ -279,27 +285,358 @@ class AbstractHowsoClient(ABC):
         })
         self._auto_persist_trainee(trainee_id)
 
-    @abstractmethod
-    def remove_cases(self, trainee_id, num_cases, *,
-                     case_indices=None, condition=None,
-                     condition_session=None, distribute_weight_feature=None,
-                     precision=None) -> int:
-        """Remove training cases from a trainee."""
+    def remove_cases(
+        self,
+        trainee_id: str,
+        num_cases: int,
+        *,
+        case_indices: t.Optional[CaseIndices] = None,
+        condition: t.Optional[Mapping] = None,
+        condition_session: t.Optional[str] = None,
+        distribute_weight_feature: t.Optional[str] = None,
+        precision: t.Optional[Precision] = None,
+    ) -> int:
+        """
+        Removes training cases from a Trainee.
 
-    @abstractmethod
-    def move_cases(self, trainee_id, num_cases, *,
-                   case_indices=None,
-                   condition=None, condition_session=None,
-                   precision=None, preserve_session_data=False,
-                   target_id=None, source_id=None,
-                   source_name_path=None, target_name_path=None) -> int:
-        """Move training cases from one trainee to another in the hierarchy."""
+        The training cases will be completely purged from the model and
+        the model will behave as if it had never been trained with them.
 
-    @abstractmethod
-    def edit_cases(self, trainee_id, feature_values, *, case_indices=None,
-                   condition=None, condition_session=None, features=None,
-                   num_cases=None, precision=None) -> int:
-        """Edit feature values for the specified cases."""
+        Parameters
+        ----------
+        trainee_id : str
+            The ID of the Trainee to remove cases from.
+        num_cases : int
+            The number of cases to remove; minimum 1 case must be removed.
+            Ignored if case_indices is specified.
+        case_indices : Sequence of tuple of {str, int}, optional
+            A list of tuples containing session ID and session training index
+            for each case to be removed.
+        condition : dict of str to object, optional
+            The condition map to select the cases to remove that meet all the
+            provided conditions. Ignored if case_indices is specified.
+
+            .. NOTE::
+                The dictionary keys are the feature name and values are one of:
+
+                    - None
+                    - A value, must match exactly.
+                    - An array of two numeric values, specifying an inclusive
+                      range. Only applicable to continuous and numeric ordinal
+                      features.
+                    - An array of string values, must match any of these values
+                      exactly. Only applicable to nominal and string ordinal
+                      features.
+
+            .. TIP::
+                Example 1 - Remove all values belonging to `feature_name`::
+
+                    criteria = {"feature_name": None}
+
+                Example 2 - Remove cases that have the value 10::
+
+                    criteria = {"feature_name": 10}
+
+                Example 3 - Remove cases that have a value in range [10, 20]::
+
+                    criteria = {"feature_name": [10, 20]}
+
+                Example 4 - Remove cases that match one of ['a', 'c', 'e']::
+
+                    condition = {"feature_name": ['a', 'c', 'e']}
+
+        condition_session : str, optional
+            If specified, ignores the condition and operates on cases for
+            the specified session id. Ignored if case_indices is specified.
+        distribute_weight_feature : str, optional
+            When specified, will distribute the removed cases' weights
+            from this feature into their neighbors.
+        precision : {"exact", "similar"}, optional
+            The precision to use when moving the cases, defaults to "exact".
+            Ignored if case_indices is specified.
+
+        Returns
+        -------
+        int
+            The number of cases removed.
+
+        Raises
+        ------
+        ValueError
+            If `num_cases` is not at least 1.
+        """
+        if num_cases < 1:
+            raise ValueError('num_cases must be a value greater than 0')
+
+        if precision is not None and precision not in self.SUPPORTED_PRECISION_VALUES:
+            warnings.warn(self.WARNING_MESSAGES['invalid_precision'])
+
+        # Convert session instance to id
+        if (
+            isinstance(condition, dict) and
+            isinstance(condition.get('.session'), Session)
+        ):
+            condition['.session'] = condition['.session'].id
+
+        if self.configuration.verbose:
+            print(f'Removing case(s) from Trainee with id: {trainee_id}')
+
+        self._auto_resolve_trainee(trainee_id)
+        result = self._execute(trainee_id, "remove_cases", {
+            "case_indices": case_indices,
+            "condition": condition,
+            "condition_session": condition_session,
+            "precision": precision,
+            "num_cases": num_cases,
+            "distribute_weight_feature": distribute_weight_feature,
+        })
+        self._auto_persist_trainee(trainee_id)
+        if not result:
+            return 0
+        return result.get('count', 0)
+
+    def move_cases(
+        self,
+        trainee_id: str,
+        num_cases: int,
+        *,
+        case_indices: t.Optional[CaseIndices] = None,
+        condition: t.Optional[Mapping] = None,
+        condition_session: t.Optional[str] = None,
+        precision: t.Optional[Precision] = None,
+        preserve_session_data: bool = False,
+        source_id: t.Optional[str] = None,
+        source_name_path: t.Optional[Collection[str]] = None,
+        target_name_path: t.Optional[Collection[str]] = None,
+        target_id: t.Optional[str] = None
+    ) -> int:
+        """
+        Moves training cases from one Trainee to another in the hierarchy.
+
+        Parameters
+        ----------
+        trainee_id : str
+            The identifier of the Trainee doing the moving.
+        num_cases : int
+            The number of cases to move; minimum 1 case must be moved.
+            Ignored if case_indices is specified.
+        case_indices : Sequence of tuple of {str, int}, optional
+            A list of tuples containing session ID and session training index
+            for each case to be removed.
+        condition : dict, optional
+            The condition map to select the cases to move that meet all the
+            provided conditions. Ignored if case_indices is specified.
+
+            .. NOTE::
+                The dictionary keys are the feature name and values are one of:
+
+                    - None
+                    - A value, must match exactly.
+                    - An array of two numeric values, specifying an inclusive
+                      range. Only applicable to continuous and numeric ordinal
+                      features.
+                    - An array of string values, must match any of these values
+                      exactly. Only applicable to nominal and string ordinal
+                      features.
+
+            .. TIP::
+                Example 1 - Move all values belonging to `feature_name`::
+
+                    criteria = {"feature_name": None}
+
+                Example 2 - Move cases that have the value 10::
+
+                    criteria = {"feature_name": 10}
+
+                Example 3 - Move cases that have a value in range [10, 20]::
+
+                    criteria = {"feature_name": [10, 20]}
+
+                Example 4 - Remove cases that match one of ['a', 'c', 'e']::
+
+                    condition = {"feature_name": ['a', 'c', 'e']}
+
+                Example 5 - Move cases using session name and index::
+
+                    criteria = {'.session':'your_session_name',
+                                '.session_index': 1}
+
+        condition_session : str, optional
+            If specified, ignores the condition and operates on cases for
+            the specified session id. Ignored if case_indices is specified.
+        precision : {"exact", "similar"}, optional
+            The precision to use when moving the cases. Options are 'exact'
+            or 'similar'. If not specified, "exact" will be used.
+            Ignored if case_indices is specified.
+        preserve_session_data : bool, default False
+            When True, will move cases without cleaning up session data.
+        source_id : str, optional
+            The source trainee unique id from which to move cases. Ignored
+            if source_name_path is specified. If neither source_name_path nor
+            source_id are specified, moves cases from the trainee itself.
+        source_name_path : list of str, optional
+            List of strings specifying the user-friendly path of the child
+            subtrainee from which to move cases.
+        target_name_path : list of str, optional
+            List of strings specifying the user-friendly path of the child
+            subtrainee to move cases to.
+        target_id : str, optional
+            The target trainee id to move the cases to. Ignored if
+            target_name_path is specified. If neither target_name_path nor
+            target_id are specified, moves cases to the trainee itself.
+
+        Returns
+        -------
+        int
+            The number of cases moved.
+        """
+        if not self.active_session:
+            raise HowsoError(self.ERROR_MESSAGES["missing_session"], code="missing_session")
+
+        if num_cases < 1:
+            raise ValueError('num_cases must be a value greater than 0')
+
+        if precision is not None and precision not in self.SUPPORTED_PRECISION_VALUES:
+            warnings.warn(self.WARNING_MESSAGES['invalid_precision'])
+
+        # Convert session instance to id
+        if (
+            isinstance(condition, dict) and
+            isinstance(condition.get('.session'), Session)
+        ):
+            condition['.session'] = condition['.session'].id
+
+        if self.configuration.verbose:
+            print(f'Moving case(s) from Trainee with id: {trainee_id}')
+
+        self._auto_resolve_trainee(trainee_id)
+        result = self._execute(trainee_id, "move_cases", {
+            "target_id": target_id,
+            "case_indices": case_indices,
+            "condition": condition,
+            "condition_session": condition_session,
+            "precision": precision,
+            "num_cases": num_cases,
+            "preserve_session_data": preserve_session_data,
+            "session": self.active_session.id,
+            "source_id": source_id,
+            "source_name_path": source_name_path,
+            "target_name_path": target_name_path
+        })
+        self._auto_persist_trainee(trainee_id)
+        if not result:
+            return 0
+        return result.get('count', 0)
+
+    def edit_cases(
+        self,
+        trainee_id: str,
+        feature_values: list[t.Any] | DataFrame,
+        *,
+        case_indices: t.Optional[CaseIndices] = None,
+        condition: t.Optional[Mapping] = None,
+        condition_session: t.Optional[str] = None,
+        features: t.Optional[Collection[str]] = None,
+        num_cases: t.Optional[int] = None,
+        precision: t.Optional[Precision] = None,
+    ) -> int:
+        """
+        Edit feature values for the specified cases.
+
+        Parameters
+        ----------
+        trainee_id : str
+            The ID of the Trainee to edit the cases of.
+        feature_values : list of object or pandas.DataFrame
+            The feature values to edit the case(s) with. If specified as a list,
+            the order corresponds with the order of the `features` parameter.
+            If specified as a DataFrame, only the first row will be used.
+        case_indices : Sequence of tuple of {str, int}, optional
+            Iterable of Sequences containing the session id and index, where index
+            is the original 0-based index of the case as it was trained into
+            the session. This explicitly specifies the cases to edit. When
+            specified, `condition` and `condition_session` are ignored.
+        condition : dict, optional
+            A condition map to select which cases to edit. Ignored when
+            `case_indices` are specified.
+
+            .. NOTE::
+                The dictionary keys are the feature name and values are one of:
+
+                    - None
+                    - A value, must match exactly.
+                    - An array of two numeric values, specifying an inclusive
+                      range. Only applicable to continuous and numeric ordinal
+                      features.
+                    - An array of string values, must match any of these values
+                      exactly. Only applicable to nominal and string ordinal
+                      features.
+
+        condition_session : str, optional
+            If specified, ignores the condition and operates on all cases for
+            the specified session.
+        features : iterable of str, optional
+            The names of the features to edit. Required when `feature_values`
+            is not specified as a DataFrame.
+        num_cases : int, default None
+            The maximum amount of cases to edit. If not specified, the limit
+            will be k cases if precision is "similar", or no limit if precision
+            is "exact".
+        precision : {"exact", "similar"}, optional
+            The precision to use when moving the cases, defaults to "exact".
+
+        Returns
+        -------
+        int
+            The number of cases modified.
+        """
+        if not self.active_session:
+            raise HowsoError(self.ERROR_MESSAGES["missing_session"], code="missing_session")
+
+        if precision is not None and precision not in self.SUPPORTED_PRECISION_VALUES:
+            warnings.warn(self.WARNING_MESSAGES['invalid_precision'])
+
+        if case_indices is not None:
+            util.validate_case_indices(case_indices)
+
+        self._auto_resolve_trainee(trainee_id)
+        cached_trainee = self.trainee_cache.get(trainee_id)
+
+        # Serialize feature_values
+        if feature_values is not None:
+            if features is None:
+                features = internals.get_features_from_data(feature_values, data_parameter='feature_values')
+            serialized_feature_values = serialize_cases(feature_values, features, cached_trainee.features)
+            if serialized_feature_values:
+                # Only a single case should be provided
+                serialized_feature_values = serialized_feature_values[0]
+        else:
+            serialized_feature_values = None
+
+        # Convert session instance to id
+        if (
+            isinstance(condition, dict) and
+            isinstance(condition.get('.session'), Session)
+        ):
+            condition['.session'] = condition['.session'].id
+
+        if self.configuration.verbose:
+            print(f'Editing case(s) in Trainee with id: {trainee_id}')
+
+        result = self._execute(trainee_id, "edit_cases", {
+            "case_indices": case_indices,
+            "condition": condition,
+            "condition_session": condition_session,
+            "features": features,
+            "feature_values": serialized_feature_values,
+            "precision": precision,
+            "num_cases": num_cases,
+            "session": self.active_session.id,
+        })
+        self._auto_persist_trainee(trainee_id)
+        if not result:
+            return 0
+        return result.get('count', 0)
 
     @abstractmethod
     def remove_series_store(self, trainee_id, series=None):
@@ -710,9 +1047,8 @@ class AbstractHowsoClient(ABC):
         if case_indices is not None:
             util.validate_case_indices(case_indices)
 
-        if isinstance(precision, str):
-            if precision not in self.SUPPORTED_PRECISION_VALUES:
-                warnings.warn(self.WARNING_MESSAGES["invalid_precision"])
+        if precision is not None and precision not in self.SUPPORTED_PRECISION_VALUES:
+            warnings.warn(self.WARNING_MESSAGES['invalid_precision'])
 
         util.validate_list_shape(features, 1, "features", "str")
         if session is None and case_indices is None:
@@ -833,7 +1169,3 @@ class AbstractHowsoClient(ABC):
     @abstractmethod
     def set_params(self, trainee_id, params):
         """Set specific hyperparameters in the trainee."""
-
-    def _resolve_trainee_id(self, trainee_id, *args, **kwargs):
-        """Resolve trainee identifier."""
-        return trainee_id
