+++ howso/direct/client.py
@@ -1430,115 +1430,6 @@ class HowsoDirectClient(AbstractHowsoClient):
 
         return needs_analyze
 
-    def remove_cases(
-        self,
-        trainee_id: str,
-        num_cases: int,
-        *,
-        case_indices: Optional[Iterable[Tuple[str, int]]] = None,
-        condition: Optional[Dict[str, object]] = None,
-        condition_session: Optional[str] = None,
-        distribute_weight_feature: Optional[str] = None,
-        precision: Optional[Literal["exact", "similar"]] = None,
-    ) -> int:
-        """
-        Removes training cases from a Trainee.
-
-        The training cases will be completely purged from the model and
-        the model will behave as if it had never been trained with them.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The ID of the Trainee to remove cases from.
-        num_cases : int
-            The number of cases to remove; minimum 1 case must be removed.
-            Ignored if case_indices is specified.
-        case_indices : list of tuples
-            A list of tuples containing session ID and session training index
-            for each case to be removed.
-        condition : dict of str to object, optional
-            The condition map to select the cases to remove that meet all the
-            provided conditions. Ignored if case_indices is specified.
-
-            .. NOTE::
-                The dictionary keys are the feature name and values are one of:
-
-                    - None
-                    - A value, must match exactly.
-                    - An array of two numeric values, specifying an inclusive
-                      range. Only applicable to continuous and numeric ordinal
-                      features.
-                    - An array of string values, must match any of these values
-                      exactly. Only applicable to nominal and string ordinal
-                      features.
-
-            .. TIP::
-                Example 1 - Remove all values belonging to `feature_name`::
-
-                    criteria = {"feature_name": None}
-
-                Example 2 - Remove cases that have the value 10::
-
-                    criteria = {"feature_name": 10}
-
-                Example 3 - Remove cases that have a value in range [10, 20]::
-
-                    criteria = {"feature_name": [10, 20]}
-
-                Example 4 - Remove cases that match one of ['a', 'c', 'e']::
-
-                    condition = {"feature_name": ['a', 'c', 'e']}
-
-        condition_session : str, optional
-            If specified, ignores the condition and operates on cases for
-            the specified session id. Ignored if case_indices is specified.
-        distribute_weight_feature : str, optional
-            When specified, will distribute the removed cases' weights
-            from this feature into their neighbors.
-        precision : {"exact", "similar"}, optional
-            The precision to use when moving the cases, defaults to "exact".
-            Ignored if case_indices is specified.
-
-        Returns
-        -------
-        int
-            The number of cases removed.
-
-        Raises
-        ------
-        ValueError
-            If `num_cases` is not at least 1.
-        """
-        self._auto_resolve_trainee(trainee_id)
-        if num_cases < 1:
-            raise ValueError('num_cases must be a value greater than 0')
-        if self.verbose:
-            print(f'Removing case(s) in trainee with id: {trainee_id}')
-
-        if isinstance(precision, str):
-            if precision not in self.SUPPORTED_PRECISION_VALUES:
-                warnings.warn(self.INCORRECT_PRECISION_VALUE_WARNING)
-
-        # Convert session instance to id
-        if (
-            isinstance(condition, dict) and
-            isinstance(condition.get('.session'), Session)
-        ):
-            condition['.session'] = condition['.session'].id
-
-        result = self.howso.remove_cases(
-            trainee_id,
-            case_indices=case_indices,
-            condition=condition,
-            condition_session=condition_session,
-            distribute_weight_feature=distribute_weight_feature,
-            num_cases=num_cases,
-            precision=precision,
-        )
-        self._auto_persist_trainee(trainee_id)
-        return result.get('count', 0)
-
     def edit_cases(
         self,
         trainee_id: str,
@@ -4428,137 +4319,6 @@ class HowsoDirectClient(AbstractHowsoClient):
                     "`context_features`."
                 )
         return context_features, contexts
-
-    def move_cases(
-        self,
-        trainee_id: str,
-        num_cases: int,
-        *,
-        case_indices: Optional[Iterable[Tuple[str, int]]] = None,
-        condition: Optional[Dict] = None,
-        condition_session: Optional[str] = None,
-        precision: Optional[Literal["exact", "similar"]] = None,
-        preserve_session_data: bool = False,
-        source_id: Optional[str] = None,
-        source_name_path: Optional[List[str]] = None,
-        target_name_path: Optional[List[str]] = None,
-        target_id: Optional[str] = None
-    ) -> int:
-        """
-        Moves training cases from one trainee to another in the hierarchy.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee doing the moving.
-        num_cases : int
-            The number of cases to move; minimum 1 case must be moved.
-            Ignored if case_indices is specified.
-        case_indices : list of tuples
-            A list of tuples containing session ID and session training index
-            for each case to be removed.
-        condition : dict, optional
-            The condition map to select the cases to move that meet all the
-            provided conditions. Ignored if case_indices is specified.
-
-            .. NOTE::
-                The dictionary keys are the feature name and values are one of:
-
-                    - None
-                    - A value, must match exactly.
-                    - An array of two numeric values, specifying an inclusive
-                      range. Only applicable to continuous and numeric ordinal
-                      features.
-                    - An array of string values, must match any of these values
-                      exactly. Only applicable to nominal and string ordinal
-                      features.
-
-            .. TIP::
-                Example 1 - Move all values belonging to `feature_name`::
-
-                    criteria = {"feature_name": None}
-
-                Example 2 - Move cases that have the value 10::
-
-                    criteria = {"feature_name": 10}
-
-                Example 3 - Move cases that have a value in range [10, 20]::
-
-                    criteria = {"feature_name": [10, 20]}
-
-                Example 4 - Remove cases that match one of ['a', 'c', 'e']::
-
-                    condition = {"feature_name": ['a', 'c', 'e']}
-
-                Example 5 - Move cases using session name and index::
-
-                    criteria = {'.session':'your_session_name',
-                                '.session_index': 1}
-
-        condition_session : str, optional
-            If specified, ignores the condition and operates on cases for
-            the specified session id. Ignored if case_indices is specified.
-        precision : {"exact", "similar"}, optional
-            The precision to use when moving the cases. Options are 'exact'
-            or 'similar'. If not specified, "exact" will be used.
-            Ignored if case_indices is specified.
-        preserve_session_data : bool, default False
-            When True, will move cases without cleaning up session data.
-        source_id : str, optional
-            The source trainee unique id from which to move cases. Ignored
-            if source_name_path is specified. If neither source_name_path nor
-            source_id are specified, moves cases from the trainee itself.
-        source_name_path : list of str, optional
-            List of strings specifying the user-friendly path of the child
-            subtrainee from which to move cases.
-        target_name_path : list of str, optional
-            List of strings specifying the user-friendly path of the child
-            subtrainee to move cases to.
-        target_id : str, optional
-            The target trainee id to move the cases to. Ignored if
-            target_name_path is specified. If neither target_name_path nor
-            target_id are specified, moves cases to the trainee itself.
-
-        Returns
-        -------
-        int
-            The number of cases moved.
-        """
-        self._auto_resolve_trainee(trainee_id)
-
-        if num_cases < 1:
-            raise ValueError('num_cases must be a value greater than 0')
-
-        if isinstance(precision, str):
-            if precision not in self.SUPPORTED_PRECISION_VALUES:
-                warnings.warn(self.INCORRECT_PRECISION_VALUE_WARNING)
-
-        if self.verbose:
-            print(f'Moving case from trainee with id: {trainee_id}')
-
-        # Convert session instance to id
-        if (
-            isinstance(condition, dict) and
-            isinstance(condition.get('.session'), Session)
-        ):
-            condition['.session'] = condition['.session'].id
-
-        result = self.howso.move_cases(
-            trainee_id,
-            target_id=target_id,
-            case_indices=case_indices,
-            condition=condition,
-            condition_session=condition_session,
-            num_cases=num_cases,
-            precision=precision,
-            preserve_session_data=preserve_session_data,
-            session=self.active_session.id,
-            source_id=source_id,
-            source_name_path=source_name_path,
-            target_name_path=target_name_path
-        )
-        self._auto_persist_trainee(trainee_id)
-        return result.get('count', 0)
 
     def get_params(
         self,
