+++ howso/direct/core.py
@@ -2536,142 +2536,6 @@ class HowsoCore:
             "weight_feature": weight_feature,
         })
 
-    def move_cases(
-        self,
-        trainee_id: str,
-        num_cases: int = 1,
-        *,
-        case_indices: Optional[Iterable[Tuple[str, int]]] = None,
-        condition: Optional[Dict] = None,
-        condition_session: Optional[str] = None,
-        distribute_weight_feature: Optional[str] = None,
-        precision: Optional[Literal["exact", "similar"]] = None,
-        preserve_session_data: bool = False,
-        session: Optional[str] = None,
-        source_id: Optional[str] = None,
-        source_name_path: Optional[List[str]] = None,
-        target_name_path: Optional[List[str]] = None,
-        target_id: Optional[str] = None
-    ) -> Dict:
-        """
-        Moves cases from one trainee to another in the hierarchy.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee doing the moving.
-        num_cases : int
-            The number of cases to move; minimum 1 case must be moved.
-            Ignored if case_indices is specified.
-        case_indices : list of tuples
-            A list of tuples containing session ID and session training index
-            for each case to be removed.
-        condition : dict, optional
-            The condition map to select the cases to move that meet all the
-            provided conditions. Ignored if case_indices is specified.
-        condition_session : str, optional
-            If specified, ignores the condition and operates on cases for
-            the specified session id. Ignored if case_indices is specified.
-        precision : {"exact", "similar"}, optional
-            The precision to use when moving the cases. Options are 'exact'
-            or 'similar'. If not specified, "exact" will be used.
-            Ignored if case_indices is specified.
-        preserve_session_data : bool, default False
-            When True, will move cases without cleaning up session data.
-        session : str, optional
-            The identifier of the Trainee session to associate the move with.
-        source_id : str, optional
-            The source trainee unique id from which to move cases. Ignored
-            if source_name_path is specified. If neither source_name_path nor
-            source_id are specified, moves cases from the trainee itself.
-        source_name_path : list of str, optional
-            List of strings specifying the user-friendly path of the child
-            subtrainee from which to move cases.
-        target_name_path : list of str, optional
-            List of strings specifying the user-friendly path of the child
-            subtrainee to move cases to.
-        target_id : str, optional
-            The target trainee id to move the cases to. Ignored if
-            target_name_path is specified. If neither target_name_path nor
-            target_id are specified, moves cases to the trainee itself.
-
-        Returns
-        -------
-        dict
-            A dictionary with key 'count' for the number of moved cases.
-        """
-        result = self._execute(trainee_id, "move_cases", {
-            "target_id": target_id,
-            "case_indices": case_indices,
-            "condition": condition,
-            "condition_session": condition_session,
-            "precision": precision,
-            "num_cases": num_cases,
-            "preserve_session_data": preserve_session_data,
-            "session": session,
-            "distribute_weight_feature": distribute_weight_feature,
-            "source_id": source_id,
-            "source_name_path": source_name_path,
-            "target_name_path": target_name_path
-        })
-        if not result:
-            return {'count': 0}
-        return result
-
-    def remove_cases(
-        self,
-        trainee_id: str,
-        num_cases: int = 1,
-        *,
-        case_indices: Optional[Iterable[Tuple[str, int]]] = None,
-        condition: Optional[Dict[str, Any]] = None,
-        condition_session: Optional[str] = None,
-        distribute_weight_feature: Optional[str] = None,
-        precision: Optional[Literal["exact", "similar"]] = None,
-    ) -> Dict:
-        """
-        Removes cases from a Trainee.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        num_cases : int
-            The number of cases to remove; minimum 1 case must be removed.
-            Ignored if case_indices is specified.
-        case_indices : list of tuples
-            A list of tuples containing session ID and session training index
-            for each case to be removed.
-        condition : dict of str to object, optional
-            The condition map to select the cases to remove that meet all the
-            provided conditions. Ignored if case_indices is specified.
-        condition_session : str, optional
-            If specified, ignores the condition and operates on cases for
-            the specified session id. Ignored if case_indices is specified.
-        distribute_weight_feature : str, optional
-            When specified, will distribute the removed cases' weights
-            from this feature into their neighbors.
-        precision : {"exact", "similar"}, optional
-            The precision to use when moving the cases, defaults to "exact".
-            Ignored if case_indices is specified.
-
-        Returns
-        -------
-        dict
-            A dictionary with key 'count' for the number of removed cases.
-        """
-        result = self._execute(trainee_id, "remove_cases", {
-            "case_indices": case_indices,
-            "condition": condition,
-            "condition_session": condition_session,
-            "precision": precision,
-            "num_cases": num_cases,
-            "distribute_weight_feature": distribute_weight_feature,
-        })
-        if not result:
-            return {'count': 0}
-        return result
-
     def edit_cases(
         self,
         trainee_id: str,
